cppbindings: Move data pipeline API into its own class
